+++ ackermann_steering_controller/test/test_ackermann_steering_controller_preceding.cpp
@@ -17,13 +17,14 @@
 #include <vector>
 
 #include "hardware_interface/types/hardware_interface_type_values.hpp"
-#include "test_tricycle_steering_controller.hpp"
-class TricycleSteeringControllerTest
-: public TricycleSteeringControllerFixture<TestableTricycleSteeringController>
+#include "test_ackermann_steering_controller.hpp"
+
+class AckermannSteeringControllerTest
+: public AckermannSteeringControllerFixture<TestableAckermannSteeringController>
 {
 };
 
-TEST_F(TricycleSteeringControllerTest, all_parameters_set_configure_success)
+TEST_F(AckermannSteeringControllerTest, all_parameters_set_configure_success)
 {
   SetUpController();
 
@@ -31,21 +32,22 @@ TEST_F(TricycleSteeringControllerTest, all_parameters_set_configure_success)
 
   ASSERT_THAT(
     controller_->params_.rear_wheels_names,
-    testing::ElementsAreArray(rear_wheels_preceeding_names_));
+    testing::ElementsAreArray(rear_wheels_preceding_names_));
   ASSERT_THAT(
     controller_->params_.front_wheels_names,
-    testing::ElementsAreArray(front_wheels_preceeding_names_));
+    testing::ElementsAreArray(front_wheels_preceding_names_));
   ASSERT_EQ(controller_->params_.front_steering, front_steering_);
   ASSERT_EQ(controller_->params_.open_loop, open_loop_);
   ASSERT_EQ(controller_->params_.velocity_rolling_window_size, velocity_rolling_window_size_);
   ASSERT_EQ(controller_->params_.position_feedback, position_feedback_);
-  ASSERT_EQ(controller_->tricycle_params_.wheelbase, wheelbase_);
-  ASSERT_EQ(controller_->tricycle_params_.front_wheels_radius, front_wheels_radius_);
-  ASSERT_EQ(controller_->tricycle_params_.rear_wheels_radius, rear_wheels_radius_);
-  ASSERT_EQ(controller_->tricycle_params_.wheel_track, wheel_track_);
+  ASSERT_EQ(controller_->ackermann_params_.wheelbase, wheelbase_);
+  ASSERT_EQ(controller_->ackermann_params_.front_wheels_radius, front_wheels_radius_);
+  ASSERT_EQ(controller_->ackermann_params_.rear_wheels_radius, rear_wheels_radius_);
+  ASSERT_EQ(controller_->ackermann_params_.front_wheel_track, front_wheel_track_);
+  ASSERT_EQ(controller_->ackermann_params_.rear_wheel_track, rear_wheel_track_);
 }
 
-TEST_F(TricycleSteeringControllerTest, check_exported_interfaces)
+TEST_F(AckermannSteeringControllerTest, check_exported_interfaces)
 {
   SetUpController();
 
@@ -55,13 +57,16 @@ TEST_F(TricycleSteeringControllerTest, check_exported_interfaces)
   ASSERT_EQ(cmd_if_conf.names.size(), joint_command_values_.size());
   EXPECT_EQ(
     cmd_if_conf.names[CMD_TRACTION_RIGHT_WHEEL],
-    preceeding_prefix_ + "/" + rear_wheels_names_[0] + "/" + traction_interface_name_);
+    preceding_prefix_ + "/" + rear_wheels_names_[0] + "/" + traction_interface_name_);
   EXPECT_EQ(
     cmd_if_conf.names[CMD_TRACTION_LEFT_WHEEL],
-    preceeding_prefix_ + "/" + rear_wheels_names_[1] + "/" + traction_interface_name_);
+    preceding_prefix_ + "/" + rear_wheels_names_[1] + "/" + traction_interface_name_);
   EXPECT_EQ(
-    cmd_if_conf.names[CMD_STEER_WHEEL],
-    preceeding_prefix_ + "/" + front_wheels_names_[0] + "/" + steering_interface_name_);
+    cmd_if_conf.names[CMD_STEER_RIGHT_WHEEL],
+    preceding_prefix_ + "/" + front_wheels_names_[0] + "/" + steering_interface_name_);
+  EXPECT_EQ(
+    cmd_if_conf.names[CMD_STEER_LEFT_WHEEL],
+    preceding_prefix_ + "/" + front_wheels_names_[1] + "/" + steering_interface_name_);
   EXPECT_EQ(cmd_if_conf.type, controller_interface::interface_configuration_type::INDIVIDUAL);
 
   auto state_if_conf = controller_->state_interface_configuration();
@@ -73,8 +78,11 @@ TEST_F(TricycleSteeringControllerTest, check_exported_interfaces)
     state_if_conf.names[STATE_TRACTION_LEFT_WHEEL],
     controller_->rear_wheels_state_names_[1] + "/" + traction_interface_name_);
   EXPECT_EQ(
-    state_if_conf.names[STATE_STEER_AXIS],
+    state_if_conf.names[STATE_STEER_RIGHT_WHEEL],
     controller_->front_wheels_state_names_[0] + "/" + steering_interface_name_);
+  EXPECT_EQ(
+    state_if_conf.names[STATE_STEER_LEFT_WHEEL],
+    controller_->front_wheels_state_names_[1] + "/" + steering_interface_name_);
   EXPECT_EQ(state_if_conf.type, controller_interface::interface_configuration_type::INDIVIDUAL);
 
   // check ref itfs
@@ -84,10 +92,10 @@ TEST_F(TricycleSteeringControllerTest, check_exported_interfaces)
   {
     const std::string ref_itf_prefix_name =
       std::string(controller_->get_node()->get_name()) + "/" + joint_reference_interfaces_[i];
-    EXPECT_EQ(reference_interfaces[i]->get_prefix_name(), ref_itf_prefix_name);
     EXPECT_EQ(
       reference_interfaces[i]->get_name(),
       ref_itf_prefix_name + "/" + hardware_interface::HW_IF_VELOCITY);
+    EXPECT_EQ(reference_interfaces[i]->get_prefix_name(), ref_itf_prefix_name);
     EXPECT_EQ(reference_interfaces[i]->get_interface_name(), hardware_interface::HW_IF_VELOCITY);
   }
 }
