+++ bicycle_steering_controller/test/test_bicycle_steering_controller.hpp
@@ -257,11 +257,10 @@ protected:
   std::vector<std::string> front_wheels_names_ = {{"steering_axis_joint"}};
   std::vector<std::string> joint_names_ = {{rear_wheels_names_[0], front_wheels_names_[0]}};
 
-  std::vector<std::string> rear_wheels_preceeding_names_ = {{"pid_controller/rear_wheel_joint"}};
-  std::vector<std::string> front_wheels_preceeding_names_ = {
-    {"pid_controller/steering_axis_joint"}};
-  std::vector<std::string> preceeding_joint_names_ = {
-    rear_wheels_preceeding_names_[0], front_wheels_preceeding_names_[0]};
+  std::vector<std::string> rear_wheels_preceding_names_ = {{"pid_controller/rear_wheel_joint"}};
+  std::vector<std::string> front_wheels_preceding_names_ = {{"pid_controller/steering_axis_joint"}};
+  std::vector<std::string> preceding_joint_names_ = {
+    rear_wheels_preceding_names_[0], front_wheels_preceding_names_[0]};
 
   double wheelbase_ = 3.24644;
   double front_wheels_radius_ = 0.45;
@@ -274,7 +273,7 @@ protected:
 
   // defined in setup
   std::string traction_interface_name_ = "";
-  std::string preceeding_prefix_ = "pid_controller";
+  std::string preceding_prefix_ = "pid_controller";
 
   std::vector<hardware_interface::StateInterface> state_itfs_;
   std::vector<hardware_interface::CommandInterface> command_itfs_;
