+++ mecanum_drive_controller/CMakeLists.txt
@@ -75,12 +75,12 @@ if(BUILD_TESTING)
   )
 
   add_rostest_with_parameters_gmock(
-    test_mecanum_drive_controller_preceeding test/test_mecanum_drive_controller_preceeding.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/test/mecanum_drive_controller_preceeding_params.yaml)
-  target_include_directories(test_mecanum_drive_controller_preceeding PRIVATE include)
-  target_link_libraries(test_mecanum_drive_controller_preceeding mecanum_drive_controller)
+    test_mecanum_drive_controller_preceding test/test_mecanum_drive_controller_preceding.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/test/mecanum_drive_controller_preceding_params.yaml)
+  target_include_directories(test_mecanum_drive_controller_preceding PRIVATE include)
+  target_link_libraries(test_mecanum_drive_controller_preceding mecanum_drive_controller)
   ament_target_dependencies(
-    test_mecanum_drive_controller_preceeding
+    test_mecanum_drive_controller_preceding
     controller_interface
     hardware_interface
   )
