+++ tricycle_steering_controller/CMakeLists.txt
@@ -72,12 +72,12 @@ if(BUILD_TESTING)
   )
 
   add_rostest_with_parameters_gmock(
-    test_tricycle_steering_controller_preceeding test/test_tricycle_steering_controller_preceeding.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/test/tricycle_steering_controller_preceeding_params.yaml)
-  target_include_directories(test_tricycle_steering_controller_preceeding PRIVATE include)
-  target_link_libraries(test_tricycle_steering_controller_preceeding tricycle_steering_controller)
+    test_tricycle_steering_controller_preceding test/test_tricycle_steering_controller_preceding.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/test/tricycle_steering_controller_preceding_params.yaml)
+  target_include_directories(test_tricycle_steering_controller_preceding PRIVATE include)
+  target_link_libraries(test_tricycle_steering_controller_preceding tricycle_steering_controller)
   ament_target_dependencies(
-    test_tricycle_steering_controller_preceeding
+    test_tricycle_steering_controller_preceding
     controller_interface
     hardware_interface
   )
